Updates to ensure that os_printf with %d types are coerced to be ints.
When os_printf is given a format of "%d" then it MUST be supplied ints.
uint32 types will not do.

+++ targets/esp8266/user_main.c
@@ -118,7 +118,7 @@ static void eventHandler(
 	// Handle the unknown event type.
 	default:
 		os_printf("user_main: eventHandler: Unknown task type: %d",
-				pEvent->sig);
+				(int)pEvent->sig);
 		break;
 	}
 } // End of eventHandler
